all except 4.16, 4.20, 4.22 and 4.23

+++ 4/blogilista/controllers/blogs.js
@@ -1,10 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
   } catch (error) {
     next(error)
@@ -27,26 +28,44 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 })
 
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  } return null
+}
+
+
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
-  const user = await User.findById(body.userId)
-  if (!body.title || !body.url) {
-    return response.status(400).json({ error: 'Title or URL missing' })
-  }
-
-  const newBlog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes === undefined ? 0 : body.likes,
-    user: user._id
-  })
-
   try {
-    const savedBlog = await newBlog.save()
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
-    response.status(201).json(savedBlog)
+    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid or no token' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+
+    if (!body.title || !body.url) {
+      return response.status(400).json({ error: 'Title or URL missing' })
+    }
+
+    const newBlog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user._id
+    })
+
+    try {
+      const savedBlog = await newBlog.save()
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
+      response.status(201).json(savedBlog)
+    } catch (error) {
+      next(error)
+    }
   } catch (error) {
     next(error)
   }
@@ -54,12 +73,24 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'Token invalid or missing' });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+    const user = await User.findById(decodedToken.id);
+
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else {
+      return response.status(403).json({ error: 'Forbidden' });
+    }
   } catch (error) {
-    next(error)
+    next(error);
   }
-})
+});
 
 blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
@@ -71,13 +102,11 @@ blogsRouter.put('/:id', async (request, response, next) => {
     likes: body.likes
   }
   try {
-    const updated = await Blog.findByIdAndUpdate(request.params.id, updatedBlog, { new: true, upsert: true })
-
+    const updated = await Blog.findByIdAndUpdate(request.params.id, updatedBlog, { new: true })
     response.status(200).json(updated)
   } catch (error) {
     next(error)
   }
 })
-
 
 module.exports = blogsRouter
